smart_home/requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
Pillow==10.0.0
pip==23.2.1
psycopg2-binary==2.9.7
pytz==2023.3
setuptools==60.2.0
sqlparse==0.4.4
typing-extensions==4.7.1
wheel==0.37.1
